// www/src/app/request/common/enum/position-history-type.ts
export enum PositionHistoryType {
  DOC_REQUEST = 'DOC_REQUEST',
  DOC_POSITION = 'DOC_POSITION',
  DOC_OFFER = 'DOC_OFFER',
  DOC_TECHPROPOSALS = 'DOC_TECHPROPOSALS',
  DOC_MESSAGE = 'DOC_MESSAGE',
  DOC_CONTRACT = 'DOC_CONTRACT',
  DOC_MANUFACTURING = 'DOC_MANUFACTURING',
  REQUEST_ADDED = 'REQUEST_ADDED',
  REQUEST_EDITED = 'REQUEST_EDITED',
  REQUEST_STATUS = 'REQUEST_STATUS',
  POSITION_ADDED = 'POSITION_ADDED',
  POSITION_EDITED = 'POSITION_EDITED',
  POSITION_STATUS = 'POSITION_STATUS',
  OFFER_ADDED = 'OFFER_ADDED',
  OFFER_EDITED = 'OFFER_EDITED',
  OFFER_STATUS = 'OFFER_STATUS',
  WINNER_ADDED = 'WINNER_ADDED',
  WINNER_REMOVED = 'WINNER_REMOVED',
  WINNER_STATUS = 'WINNER_STATUS',
  CONTRACT_ADDED = 'CONTRACT_ADDED',
  CONTRACT_EDITED = 'CONTRACT_EDITED',
  CONTRACT_STATUS = 'CONTRACT_STATUS',
  POSITION_TECHNICAL_PROPOSAL_STATUS = 'POSITION_TECHNICAL_PROPOSAL_STATUS',
  DESIGN_STATUS = 'DESIGN_STATUS'
}